Add files via upload
The project is based on real-life applications of the data structures heaps and red black trees. This project devises an effective mechanism for building-construction of a city with multiple buildings. Red-black trees and Min-heaps are used to monitor current progress, time taken, and total time required for completion of said constraints of the
city efficiently with an objective to minimize space and time complexity.