README.md
This is a project focused on the applications of data structures in the real-world applications. This project devises an effective mechanism for building construction of a city with multiple buildings. Red-black trees and Min-heaps are implemented to monitor "current progress", "time taken", and "total time required for completion" of said constraints of the city efficiently with an objective to minimize space and time complexity.